Implemented fog and removed the skybox coloring
Also changed the behaviour with how the tree is made invisible.
The game object is now being disabled entirely, instead of just disabling the renderer component.
This should fix raycasts from for example the teleporting-mechanic to hit the trees.

The "MySkybox" material that was needed for the coloring was also removed from the project.